DEV-667 Fix issue with cluster scaling with no-upload
Fixes a bug in the check for "gz"'d files. The old check would just take
the first the file in the list, which turned to be the "/samples" dir
itself.

New version actually checks file by file, so will also support dirs that
have a mix of zipped and unzipped files.