DEV-658 Use scheduled deletion and remove the janitor
Google has released a new feature in beta which makes the janitor
obsolete. It allows for specifying an "idel time to live", which will
delete a cluster if no job has been submitted for a specified ttl. This
is exactly what we want.

In order to use the feature, we must use the beta version of the api.
This is a risk, but this feature should not change much beyond perhaps
moving the parameters around the datamodel. Rolling it back is trivial
so, we're ok with that risk.

The idleTtl can be passed as an argument to bootstrap, defaulting to
the minimum of 10 minutes (600s).

We also delete the janitor with this commit. Less code.