DEV-665 Add BQSR to pipeline 5
To evaluate how well ADAM's bqsr matches GATK we need to add it to the
pipeline. Runs BQSR as a separate enrichment stage.

Removes the "output type" and surrounding complexity of supporting
persistence of sub-stages of the BAM creation.

BQSR takes a new config as input, known snps. It combines these with the
known indels, as that is what we do now in production.

A functional test is added that checks that the average difference
between GATK and ADAM per read never exceeds 50%. This is arbitrary at
this point and hope we can add something better.

DEV-665 Add BQSR to pipeline 5

To evaluate how well ADAM's bqsr matches GATK we need to add it to the
pipeline. Runs BQSR as a separate enrichment stage.

Removes the "output type" and surrounding complexity of supporting
persistence of sub-stages of the BAM creation.

BQSR takes a new config as input, known snps. It combines these with the
known indels, as that is what we do now in production.

A functional test is added that checks that the average difference
between GATK and ADAM per read never exceeds 50%. This is arbitrary at
this point and hope we can add something better.