DEV-665 Validation of BQSR
Split the output BAM into a base recalibrated and non-recalibrated. The
former will be used by strelka, the latter by the rest of the pipeline.

Also fixed a bug in which the known snp sites were not being used in
the recalibration.

Adapts the composer to also run over the recalibrated BAM so it can
be downloaded later. Note it is not yet added to the sorting or the
download, which will need some rethinking as a result of this change.

Rename fa files to fasta to make them a little more standard with other
tools