DEV-449: Execute processes on ComputeEngine VMs
Up until now we've been running all GCP jobs on DataProc clusters, which
works most of the time but which isn't a great fit for all applications.
This commit implements the ability to run arbitrary applications on
ComputeEngine VMs. It includes:

  * Scripts for building "base" images for each application that we
    decide to run in ComputeEngine;
  * Code for bootstrapping, managing and stopping VM instances;
  * The ability to run arbitrary applications on VM instances and some
    basic log and process management.

Also included is everything needed to run GATK's HaplotypeCaller
application in a VM and a code module that is used to pull down the GATK
dependencies in order to produce a shadowed JAR.