DEV-682 Split Bootstrap into main, provider, and actual bootstrap
Bootstrap had three responsibilities: running the steps to produce an
alignment on GCP, deciding whether we were running locally or in SBP and
wiring up the right dependencies, and being an entry point. This commit
splits up these roles for readability, testability and extensibility.

The provider handles the wiring. The name provider itself is a nod to
javax.inject.Provider, but not using that interface as there is no DI
framework here. To remove the if statement we use an abstract factory
pattern. A simple unit test is added to ensure no errors in wiring and
that the right implementation is used when the sbp id is in the args.

Main was split out as it looked a bit strange to have Bootstrap depend
on its provider which depends on Bootstrap. This will make it easier to
use Bootstrap as a component ina larger pipeline later.

Also renamed static data to resources, but unrelated but wanted to do
that for awhile :)