DEV-679 Introduce argument profiles and remove all unnecessary args
This change adds the notion of argument profiles, to default well both
locally and in production. The development profile relies on a few
conventions but aims to reduce args down to just the sample id.

Conventions:
- Sample exists in the /samples root of the project
- Private key is called bootstrap-key and is also in root of project
(it is ignored by git)
- Running on a mac (sorry Ned)

All profile related defaults are in the Arguments class.

BootstrapOptions is simplified by removing all the defaults, and a
bunch of dubious options.

Lastly when running in development mode with no runid specified, a new
one will be genereated each time as the timestamp.