DEV-687 Respect boolean defaults in options
Fixes a bug in which the boolean defaults were ignored when constructing
the final Arguments.

To add the ability to override defaults of true, we need someway to
specify false (old behaviour was that if the flag was present it was
true). So we now have to pass true|false to these args. Couldn't figure
out a way with apache CLI to have cake and eat it to.

Any non-boolean to these args will throw a parse exception.