DEV-681 Refactor cluster interface to be more executor-like
Originally the cluster had a start/submit/stop interface with the idea
that a job could be submitted to the same cluster multiple times. In
practice we don't do that and that setup made it awkward to make start-
ing the cluster depend on whether there were any jobs left to be exec-
uted.

We replace that interface with an Executor-like equivalent, with only
submit. Submit will be charged with also starting and stopping the
cluster (if required). This actually simplifies quite a bit around it,
for instance we don't need two "clusters" in bootstrap, just a single
executor, which is now more of an abstraction for Dataproc itself than
a single cluster within Dataproc.

Since starting and submitting are now a single operation, its easy to
check if the job has completed before starting the cluster, fixing this
issue.

This will also fit nicely with pending changes to support execution on
standalone VMs.