DEV-696 Preperation work for adding somatic calling
To this point the entire "cluster" module (the google pipeline) has
been focused on BAM creation and enrichment of a single sample (ie a
single set of fastq).

For strelka, we'll need both the reference and tumour samples so we need
to draw a distinction. To this end we do the following refactoring:
- Rename bootstrap to alignment, bootstrap was meant only for setup but
evolved into the alignment sub-pipeline.
- Move CLI related stuff up a package level
- Introduce PatientReportPipeline, which will tie the sub-pipelines
together.
- Introduce skeleton caller sub-pipelines

The idea is that each sub-pipeline (alignment, germline calling, somatic
calling, structural calling and eventually tertiary) will do some work
and return a pointer to the output data in google storage to be consumed
by the next stage.

Tricky bit is how to "schedule" the full run when each sample is run
asynchronously. The idea here is that whenever an alignment/germline
call finishes we check to see if the complementary sample exists (ie if
I'm a tumour, is there a reference bam? and vice-versa). If so, we kick
off variant calling.

Note the code is current all serial, we'll think about concurrent
processing later.