DEV-696 Add functional alignment storage
The role of the alignment storage is to provide pointers to existing
data in GS. This is used in two places: after the alignment has run to
ensure it has really run successfully; when looking for the complement
sample (ie the tumor to the ref or vice-versa).

Shuffled around some dependency injection to allow for single storage
and credentials to be used in multiple contexts.

Removed JobResult from the AlignmentOutput, I think we'll be improving
result reporting in the near future so no sense adding more dependencies
to the current way of doing things.

Then just filled in any nulls/gaps to make the pipeline actually run as
is.