DEV-698 Refactor initial germline/vm code to share with strelka
Goal of this refactor is to share more concepts and code with the
alignment stage in order to make it easier to add strelka without
duplication.

All vm and dataproc code moved into a shared package structure and
decoupled from pipeline steps.

Conceptually we move from the GoogleVirtualMachine abstraction to the
ComputeEngine abstraction, as we did with SparkCluster moving to
Dataproc. Constructor arguments are moved into the submit method, passed
via a job definition object.

Wiring moved to the provider classes and shared dependencies (storage,
credentials, etc) passed to both vms and dataprocs from patient report
pipeline.

The runtime bucket is also now used in the germline calling.

Fixed all intellij warnings, including the use of optionals in the
bash builder, which was a little more intrusive.

Removed custom exception types, as we have decided not to make use of
custom exceptions in this codebase.