DEV-698 Add the strelka vm and input/output data
A little more refactoring as well, moved the performance code into the
execution packages and drew a clearer distinction between what's
relevant for dataproc and compute engine. There is now a shared marker
interface.

Strelka starts to come together. The image commands are added to the
scripts directory. The strelka image copies our strelka binaries from
storage and creates some directories for inputs and resources.

RuntimeBucket is extended to also work for a pair of T/R samples.

SomaticCaller itself now creates the right VM and copies the input files
into a /data/input, which can be our general convention.