DEV-698 Copy resources into runtime bucket for VMs
Wires a Resource into somatic caller rather than just using the
argument string. This allows us to copy the resource into the runtime
bucket as we do in alignment, ensuring we know exactly which ref genome
was used for the run.

We then add a bash command to download the resource from gs using the
resources target bucket.