DEV-698 Wire input data into configure strelka workflow
Takes the last commit a bit further, exposing the locations of download
ed resources to be used in the strelka command.

InputDownload is introduced as a new command that downloads input and
keeps the end local path as state.

Resources now return a ResourceLocation after copy, which can also be
used to reference the local path on the vm.

The somatic caller now functions up to the point where configure
workflow is called.