DEV-698 Copy output to persistent bucket
Adds the OutputUpload command to copy everything from the scripts
output directory up into the results folder of the runtime bucket. Also
adds a JobComplete command to share what's been done for GermlineCaller.

The compute engine now checks for the completion flag in any sub-path of
the output bucket.

All vcfs are now returned from somatic caller, in the future we may run
the merging in the caller itself and the output signature will change.

Also added flags to turn the aligner and callers on and off, using
persistent data if available.d