DEV-705 Combine VCFs after strelka
First step in post-process. For now it happens on the same VM as the
strelka job itself. Mainly to keep things simple for now.

CombineVCfs is another GATK command. Added some general code to remove
duplication between the haplotype caller and this new command, and to
make it easier to call GATK in the future.

GATK is added to the strelka image, although the next step will be to
create a generic "pipeline5" image which all VM jobs will use (which
will have all toolks installed.)

Lastly, a new alias is added to handle GATK's issue with dict files.
It prefers that dict is the only extension of the file.