DEV-449: Add failure handling for VM-based jobs
The generated shell script that gets pushed to the instance at startup
now contains better error handling so we don't rely on a successful
run to push the expected "completion flag" to the bucket. The flag will
now get created by the startup script if any stage of the process (data
copying, the job itself, any interactions with storage) fails. Also add
the VmDirectories interface and remove some hardcoding of paths as well
as standardise the location of the output log for the stage.

Also update the Germline caller to fit into the updated API for copying
reference data and passing aruments so it will work with the new patient
pipeline code. There are still some pieces that are hardcoded in the
Germline caller which need to be completed but this is much closer to
where it should be.