DEV-705 Break down somatic caller into sub-stages
The first pass at the somatic caller and post-process was pretty messy,
with a large block of logic, lots of subtle duplication and inconsist-
ency in file naming conventions.

This change breaks down the caller into substages. Each substage can
be defined as one or more commands which take an input file and produce
an output file for other stages.

The stages are strung together using java.util.Function composition (and
Then).