DEV-720 Run callers in parallel
Uses a fixed thread pool executor of 3 to run the callers in sep-
arate threads. Futures are used to wait for the result. Due to the
optionality of running the callers (both if this is a single sample run
and when they are explicitly turned off), we must wrap the futures in
optionals.

All status are checked after the futures return. At this point we're
only logging failures, not failing hard (see DEV-719)