DEV-691 Move BamMetrics from Dataproc to ComputeEngine
The BAM metrics job is more amenable to our ComputeEngine flow now than
Dataproc.

Also add a "skipped" job status and modify the ComputeEngine class to
mark a job as skipped and return without doing any work if evidence of
a previous run is detected in the bucket at startup.