DEV-691: Cleave VM performance profiles from Dataproc
The ComputeEngine was sharing performance profile definitions for the
VMs with the DataProc code but they work slightly differently. The code
was giving the impression the RAM and CPU was being set by the profile
but for the VMs it was not.

As part of this make it easy to specify the number of cores and amount
of RAM explicitly; this makes sense for the VM-based loads where it may
not for the Spark-based ones.