DEV-721 Add Amber as a vm job
Fairly straightforward to add the remove VM job, Amber is a single
command requiring the ref genome, a pon bed, and both bams as input.

Made a couple of common test classes to share with SomaticCallerTest.

Trickiest bit is getting R and its dependencies set up. We install R
version 3.6 (required by Amber) and then add an installer script for
the dependencies. The installer script lives in google storage but I've
added to the resources so we have it versioned somewhere.