DEV-727 Refactor "ResultsDirectory" to support namespaces
The ResultsDirectory class was already used in most places to wrap any
place where we persists results, so to add namespacing we can just
extend this class with a new parameter.

That said it wasn't used everywhere. Bam creation also had "results/"
hardcoded so we needed to pass it as a parameter to the job. Also a
bunch of tests needed to "namespace" paths.