DEV-727 Use runtime bucket as namespace instead of results directory
The prior commit to this ticket was a bit short-sighted, the results
all end up separated but there are still a lot of collisions around
resources, job status, etc.

Using the top level concept, the runtime bucket makes everything much
tidier. When a user enteres the runtime bucket they will now see the
namespaces at the top level, which is nice and intuitive.

Implementation was to move all interaction with buckets into the runtime
bucket class, adding get, create, list, copy and compose operations.
Was still a bit fiddly even after this, so still room for improvement
here.

One dubious "just get it working" decision is to use the namespace in
the "bucket" field of GoogleStorageLocation, needs more thought.