DEV-725 First purple integration
Adds the purple stage to the pipeline. Purple needs input from somatic,
structural callers, cobalt and amber. The last two provide their input
to purple as directories, so some changes required to InputDownload to
allow for copying of whole directories.

Purple itself is a straightforward call to a Java program. At this point
it only returns a status, but that may change when we add the health
check.  As the structural caller is not yet complete, purple will not
function with this commit.

PatientReportPipeline has become a bit awkward tying together all the
optionals to a single point. That will be address in a subsequent
commit.