DEV-731 Add health checker as a vm job
The final step in patient report pipeline, health checker requires input
from bam metrics, somatic caller, amber and cobalt.

Also includes a refactoring of bam metrics to be a full fledged stage
in the pipeline so we can use its output in the health checker. We also
require the mate's metrics, so added a storage class to retrieve them.
At this point they do not actually check if metrics exist, this is
assumed as we know at this point the mate bams exist. Plan is to
refactor this further and check properly in a later commit.

Propagating the purple output directory in normal way.