Simplify and test PatientReportPipeline
Was getting pretty hairy with all the conditionals, optionals and
futures. Also we weren't unit testing anything or handling failures from
the stages.

First we remove all the argument checking from the pipeline and move it
into the individual stages. They now all return a SKIPPED status if they
are disabled in the args.

This gets rid of quite a few of the optionals. Removes the functional
use of Optional.map in favor of a simple isPresent check, in this case
looks more readable this way.

A state object is introduced to keep track of outputs as the jobs return
asynchronously.  The state object is checked at various important points
throughout the pipeline (after aligner, after callers, after purple) in
order to fail hard when subsequent steps won't have their output