DEV-734 Consolidate resources into one bucket
Refactors the Resource class to use a single bucket with subdirectories
rather than a bucket per resource.

Arguments change to match this approach, instead of having the ability
to use a different bucket for a couple of the resources, you can now
point all resources to another bucket. This will be useful to avoid
naming conflicts between production and dev with one parameter.

Names of resources are consolidated into a constants interface.

The Resource class now accepts the bucket and a resource name in order
to resolve the resource. Also allowed some simplification of the naming
as a sub dir can have underscores and need not be unique across
projects.