DEV-740 Centralize versions and use everywhere
Creates a Versions interface, a simple group of constants containing all
versions of tools we use. Perhaps as a next step we'll add a proper
interface that versioned tools must conform, but for now this at least
makes it easy to see each version.

Bigger change is to use gs://common-tools in node-init.sh, so that we
can also specify bwa and sambamba in code.

Removed the travis tag command as it didn't work due to permissions.