DEV-691 Fix first cohort of problems discovered during manual testing
In particular some files are produced as part of the intermediate steps
that are expected to be in certain places by later steps so add handling
for this.

Add the required software to the standard image and pull in the config
and blacklist files from the current pipeline.

Also refactor both the test layer and the implementation of the commands
to improve reusability, taking advantage of the fact that many of the
called commands share common command lines.