DEV-739 Start output with single sample logs
This commit is the beginning of compiling the final patient report.
The patient report is comprised of components which can be added to
the individual stages. This means that files will only be added to the
patient report when that step is actually run, and not when skipped.

Also introduces a simple abstraction for a metadata api service, this
will eventually use th SBP API to find the "set name", but now just
assigns a unique set name based on run id or timestamp

Getting the dataproc logs is a little tricky, we need to scan the root
of the runtime bucket for dirs prefixed with the metadata root, then
parse out the driver logs.