DEV-691 Add Filter stage for GRIDSS
This stage is not yet incorporated into the structural caller flow
because we are having some difficulties with the dependencies on the R
side in the GCP VMs. However when they are sorted out we should be
able to enable the stage and thus complete the GRIDSS pipeline.

Also rename the output VCFs in the structural output to align with the
GRIDSS terminology rather than the terminology used in the later steps.

Add some dependencies to the standard image that are needed to run the
GRIDSS R script.