DEV-760 Cleanup all runtime buckets on success
Adds a cleanup stage at the end of the "somatic" pipeline (when the
condition that the other sample in the set is also complete) which will
delete runtime buckets based on the reference and tumor sample names
and the run id argument.

It's all a little awkward without a clear distinction between the single
sample and somatic pipelines, plan to improve this in a future refactor.