DEV-768 Share code to use all CPUs as an argument.
Was being done in a few different places so moved it into a utility
function.

Also include the -nct for all GATK commands. They don't all support it
but it doesn't do any harm for those tools. May be some memory tuning
to do as a result (more threads = more heap).