DEV-770 Allow for no sample id in arguments
SBP will only pass the integer sbp_sample_id so we need to loosen the
restriction on sample_id being mandatory.

This also means a little reworking of the PatientMetadataApi, which is
really just a placeholder for now. Currently will just return the sbp
id and a dummy set name when called in SBP mode.