DEV-768 Speed up WGS metrics collection
Switch to using the GRIDSS jar to do the metrics collection as it can
decompress and parse the inputs in parallel.

Also remove the dependency between SV caller and bam metrics, as they
are not actually used in GRIDSS. This means we do not have to wait for
metrics before starting all callers, only before we do the health
check.