DEV-773 Split single sample and somatic pipelines
Creates a new pipeline called SomaticPipeline containing all stages
using both samples. These steps are removed from PatientReportPipeline
which is renamed to SingleSamplePipeline. A new class PipelineMain is
created as the runnable entry point, and makes the decision of which
flavor to run.

This is decided by a new argument, mode, which has two possible values:
single_sample; somatic. The former works identically to the old pipeline
; the latter needs an additional argument, set_id, in order to work.
The set_id for a local run should just be the name of the "patient", ie
CPCT12345678. The set_id for SBP should map to a valid integer set id
in the API.