DEV-775 Store gsutil cache on tmpfs
Add an `/etc/fstab` entry to mount a `tmpfs` on `/root/.gsutil` so that
when the VMs boot they start with nothing in that directory. This is one
way to avoid the 403 errors we see intermittently after a new disk
image is published. Saves us polluting our application code with what
amounts to an infrastructure workaround.