DEV-782 Update SBP API with status
Fleshes out all SBP interaction including getting the initial sample and
set information from the api (was previously a dummy value), updating
the sample status at the end of the single sample pipeline and updating
the run status at the end of the somatic pipeline.

An additional lookup by sample id is added so that we use the correct
set name in the single sample pipeline. If no set exists for a sample
we throw a runtime exception.

When the alignment is complete we update the api for the sample, as from
that point on somatic calling can begin.

The same update logic is applied to the runs endpoint after the health
check. Status can be Failed or SnpCheck, the latter means success from
our end, and that the SnpCheck still needs to be run.

We still need to add logic to verify that the health check is completed
successfully.