DEV-782 Refactor to use barcodes as sample ids
An extensive refactoring to decouple a samples unique id from its name,
therefore enabling us to use barcode as the sample id for SBP runs.

The barcode will now be used in all bucketing of output, although the
files themselves will still use the sample name for familiarity.

Single sample and somatic pipelines now share the concept of run meta
data (the former just the sample name and id, the latter the run/set id
along with two single sample run metadatas).

The metadata objects are used in all RuntimeBucket creation.

The somatic metadata is extracted from the SBP api by querying for the
set from the run, and the samples from the set. That way we get the
exact barcodes for this run.

Locally sample id and name are always the same.