Remove all trace of bam download
In Pv5 hybrid setup we used to download the BAMs at the end of alignment
to S3 or locally depending on setttings. We now only want to do a single
transfer to S3 on somatic pipeline completion, and leave it to the user
to decide to download BAMs for local research.

To this end, this commit removes all the code used to do the transfer
and update S3 metadata.