DEV-691 Switch to GRIDSS 2.4.0
After much testing and tuning we've switched to 2.4.0 because it is the
only version that we've run which has successfully finished when run
against the Pv5 BAMs. There is still quite a lot of performance we
think we can eek out of the structural calling but this implementation
should work which is the first important step.

Also add the "-n" flag to the `gsutil` command for input downloads. This
addition should have no impact for regular production runs but is
invaluable when attempting to do iterative improvements on a VM in
production. The flag makes `gsutil` behave with a "no clobber" attitude
so that restarting a VM manually doesn't result in all the input
files being re-downloaded.