Fix issue in filtering on final bucket
Since we were using only a prefix, other runs could be included in the
transfer. For instance if our new run has the cancerPanel/ root, we'd
also pick up cancerPanel-my-test-run/. Adding a trailing slash fixes
that.

Also logging an error and continuing on the null md5 as we investigate
the issue.