DEV-776 Fix issues with null md5 on dataproc logs
For some strange reason, when we compose the ..0000 and ..00001 dataproc
logs they end up with null md5s. There is something in the docs about
the compose operation not doing an md5, but we've verified it works for
the BAM files, which are also composed.

Since the 0001 file is always empty, and there is no good reason for the
log files to roll here (other than a disaster, which would be visible in
the first log). We'll just copy the 0000 log and use it as our run log.