DEV-803 Eliminate the need to download the SnpEff DB from the Internet
The database was copied to the resources bucket and is now pulled down
by the germline caller upon startup. Note that we have not explicitly
disabled the download in the SnpEff configuration, but it determines
it does not need to do it on startup. It should no longer be able to
do a download in production anyway as the networking will be too
restrictive.

Also change some INFO log statements to DEBUG in various places.