DEV-807 Automate the promotion of config from dev to prod
Very simple script to just promote the two buckets (tools and resources)
by deleting the production one, creating a new bucket, and copying the
dev versions.

Also modified the image script to take project and tools bucket as args
in order to run it against the production project without changes to
the cmds.