DEV-776 Use gsutil for results upload to SBP
We've found a strange issue with rclone where it tries to infer the mime
type from the file extension. This occassionally goes wrong with the SBP
object store. From what we can see there is no "fix" for this or option
to turn this behaviour off.

So, we'll try it with gsutil. May revert this later if we find a way
with rclone.