DEV-804 Refactor GRIDSS code before switch to functional style
Many utility methods and static classes have sprung up in the
GRIDSS code owing as much to the imperative style as to a desire
to box all untested code into a couple of small classes. It is
desirable to make the GRIDSS execution code look more like that
of the somatic and germline pipelines.

Remove most static methods in favour of constructor injection,
switch to inheritance-based implementation of commands, remove
GridssArguments, GridssToBashCommandConverter, GridssCommon,
replace "one-off" commands with classes.