DEV-841 Add smoke test for docker profile
Will run a simple smoke test consisting of starting the full pipeline
and asserting that the pipeline succeeeds.

Credentials are passed in via an encrypted private key file. The file
is encrypted with travis who also holds the decryption key.

A junit Category is used to separate this test from the rest. Currently
disabled to test running in travis as part of the PR build.