DEV-846 Refactor GRIDSS subsystem
GRIDSS was the odd duck of the callers as it had not been written to
conform to the SubStage setup that was used elsewhere. It has now been
rewritten to mostly look like the others.

Also all the BashCommand subclasses have been modified to use constructor
injection as is used elsewhere rather than taking parameters via method
calls. This was necessary to take advantage of the SubStage structure and
has the nice benefit of simplifying testing. The CommandFactory and
related classes have been removed.

Also some tests have been added for the StructuralCaller though there are
more to write still.

Not quite finished yet because I have not yet written a test for the
structural caller. That will be done before this branch is pushed.

Other changes were to make it so the PipeCommand always wraps its elements
in a subshell. A later commit to this branch will ensure all existing
usages do not double up on the subshells as they have often been used
together.