DEV-845 Combine system and bam modules
Merge all classes, tests and resources into the BAM module.

Added a bam package to the bam module to make it a little more clear
that these classes are only for bam creation.

Also duplicated the Processes class to avoid having the cluster module
need to bring in all the hadoop common dependencies.