DEV-851 Retry delete if the job still exists
Grasping at straws a little as we cannot duplicate this issue with some
jobs not getting deleted. Assuming that its a Google issue, this commit
will retry the delete if the job still exists.

Using failsafe again to implement the retry mech. The failure condition
is a non-null return value of jobs.get. On failed attempt we try to
delete the job again.

Annoyingly, the google api returns a 404 exception instead of a response
object when get doesn't find anything, so we have to catch that and
return null to trigger the "success" scenario.