DEV-853 Empty the metadata of the instance on failure
This should stop the startup script from executing once the vm has
failed.

The startupscript is disabled by clearing the metadata of the instance.
Note the google api checks the hash (fingerprint) of the metadata
before update, so you need to know the actual hash set by the compute
engine before updating. This is why we do the get before update.