DEV-854 Simple bcl2fastq application using compute engine
First pass, new application which runs bcl2fastq. Functionally it does
the following:
- Copy down the bcl file by the flowcell id
- Run bcl2fastq
- Copy all fastq back to a runtime bucket for the flowcell

In the future it'll have to distribute the different patients to their
respective runtime buckets.

Some awkward reuse of arguments, in the end we'll either have to
continue with a monilith argument class or extract the common ones into
a shared interface. Probably a few things we'll want to make more
shareable package-wise (or maybe separate modules).