DEV-846 Tidy up Filter stage
An Rscript command has been extracted instead of the inline
BASH code generation in previous versions. Also the Filter was
changed to copy its input to its output, which I'm not really
sure about but it one way to deal with the fact that the input
is also used by the report composition.

Also some redundant parameters were removed from the Annotation
and Calling stages and command generation was made more readable
in the stages by inlining variables and chaining additions to
the Bash commands.