DEV-868 Support single sample sets in somatic pipeline
As it fits better within SBP automation, we've decided to do the single
sample file transfer and cleanup by skipping the somatic algos in the
somatic pipeline when the SBP API returns no tumor sample in the set.

Requires changes to the somatic pipeline to skip the steps, the
cleanup to not attempt to cleanup a non-existent tumor run, and the
api interaction to allow for a null tumor.