DEV-867 Replace underscores and truncate long barcodes
In production we've observed that barcodes can occassionally be suffixed
with some information about the study id (ie FR16672550_CPCT02050443t
is a sample which is part of both CPCT and WIDE, and this is the CPCT
one. This causes two issues: underscores are not allowed in resource
names; longer run ids than Google supports.

This change truncates the barcode/sample ids to 13 to ensure we have
enough slack to stay under 50 chars with the job name.

It also replaces any underscores with dashes.