DEV-869 Use SBP API to find whether a sample has dependent runs
Before deleting it. This will ensure we never delete data still in use
by a processing run.

This is done by querying for runs by sample name, unfortunately I can
find no way of linking a run to a sample by barcode or sample sbp id.

We always delete the somatic runtime bucket as it is directly tied to
this run.

Note we reserve the right to move this to a separate cleanup process in
the future, but this will be the fastest time to market for this fix.

Also note reverted change to ratio back to 4 as our testing today
showed no benefit.