DEV-873 Wait for single sample results when staging final pipeline
Adds an infinite retry when composing the full somatic results for both
the reference and tumor pipeline outputs. Infinite because it can be
hours later and we should detect any hanging samples elsewhere.

Also removes the deletion of the single sample staging buckets. This
will allow for reuse of the same single sample run in multiple sets,even
if the other sets have not run yet. Although we will have to find a way
to clean these up in release 5.2.