DEV-868 Fix issues with retrieving ini name
Our mock run json was old and doesn't reflect how the api looks now.
There is no ini name in the run, only the id reference to the inis api.

This change calls the inis api to check the ini name. The ini api does
not seem to adhere to resource conventions so we query for all inis and
find the correct one by hand.

Tests adjusted accordingly.