DEV-860 Use SSD local scratch disks for /data
The pricing for local "scratch" SSDs is quite a bit
less than that of persistent SSDs, or even persistent
magnetic disks. This is predominantly due to the
ephemeral nature of these disks, but that suits our
workload perfectly.

The management is not as seamless for the local SSDs.
They are exposed as multiple devices so this commit
includes code to make a RAID-0 device out of them
(we need more than one as they are 375GB each) and
mount it to /data, and create some of the directories
that we used to put in the imaging script.

Further I have moved the tools out of `/data/tools`
and into `/opt/tools` as it is available at image
creation time, which is when we copy down all the
tools to the image.

The necessary code has been added to the compute
engine classes and the root disk image size dropped
to 10GB because all the space consumption for these
machines is on /data.

Note that the changes to the image introduced by
this commit make the image incompatible with any
version of the code that continues to use
`/data/tools` instead of the new `/opt/tools`.