5.3 Validation fixes in GRIDSS
Some issues with the file naming. The "unfiltered" VCF became "filtered"
"final" somewhere in the refactoring for DEV-846. We now use the the
annotated vcf and alias it with "unfiltered" when copying to report
bucket.

Still quite a bit to improve here, but keeping it small to get this
release out the door.