DEV-900 Add completion flag for output staging
Simply writes a file called STAGED when the output is complete. Keeping
it very simple for now, and not drawing a SUCCESS and FAILURE
distinction. Now if the staging process fails the somatic pipeline will
just keep waiting, which I think is nice as we may be able to correct
the issue with the upstream pipeline.