DEV-901 Enable pre-emptible vms and restart them if they are terminated
Pass the pre-emptible vm flag through to the compute engine.

When a pre-emptible vm is terminated, the instance will have the
TERMINATED status. We now query for this, and when encountered will
move on to the next zone. If we have cycled through all zones we'll
start again at the first.