DEV-905 Separate shallow runtimes and always cleanup
Solves the issue of accidental reuse of shallow data by deleting the
shallow buckets even if a dependency is detected on the sample by
another run. The dependency is likely from a different sequencing.

Also appends 'shallow' to each shallow runtime resource. Not strictly
required given the above, but should make it a little easier to see on
first inspection what is running.